fix: added better support for BaseAdapterOptions
